move export of type Newsletter
