+++ apps-rendering/src/bodyElement.ts
@@ -5,6 +5,7 @@ import type { TimelineEvent } from '@guardian/atoms-rendering/dist/types/types';
 import type { Atoms } from '@guardian/content-api-models/v1/atoms';
 import type { BlockElement } from '@guardian/content-api-models/v1/blockElement';
 import { ElementType } from '@guardian/content-api-models/v1/elementType';
+import type { ArticleTheme } from '@guardian/libs';
 import type { Option, Result } from '@guardian/types';
 import {
 	err,
@@ -24,7 +25,6 @@ import { parseImage } from 'image';
 import { compose, pipe } from 'lib';
 import type { Context } from 'parserContext';
 import type { KnowledgeQuizAtom, PersonalityQuizAtom } from 'quizAtom';
-import { Newsletter } from 'item';
 
 // ----- Types ----- //
 
@@ -111,8 +111,14 @@ interface AudioAtom {
 	title: string;
 }
 
-interface NewsletterSignUp  extends Newsletter {
+interface NewsletterSignUp {
 	kind: ElementKind.NewsletterSignUp;
+	id: string;
+	displayName: string;
+	frequency: string;
+	description: string;
+	group: string;
+	theme: ArticleTheme;
 }
 
 type BodyElement =
