+++ apps-rendering/src/server/editionsPage.tsx
@@ -16,7 +16,7 @@ import { atomCss, atomScript } from 'components/InteractiveAtom';
 import Meta from 'components/Meta';
 import Scripts from 'components/Scripts';
 import type { Response } from 'express';
-import type { Item, Newsletter } from 'item';
+import type { Item } from 'item';
 import { fromCapi } from 'item';
 import { JSDOM } from 'jsdom';
 import { compose } from 'lib';
@@ -31,6 +31,7 @@ import {
 } from 'styles';
 import { Stage } from './appIdentity';
 import { insertNewsletterIntoItem } from './insertNewsletter';
+import type { Newsletter } from './newsletterData';
 
 // ----- Types ----- //
 
@@ -158,7 +159,7 @@ const buildHtml = (
 // TO DO - update @guardian/apps-render-api-models, remove the hack on the RenderingRequest type
 function render(
 	imageSalt: string,
-	request: RenderingRequest & {promotedNewsletter?: Newsletter},
+	request: RenderingRequest & { promotedNewsletter?: Newsletter },
 	res: Response,
 	getAssetLocation: (assetName: string) => string,
 	themeOverride: Option<ArticleTheme>,
@@ -168,7 +169,7 @@ function render(
 	const environment = getEditionsEnv(isPreview, path);
 	const item = fromCapi({ docParser, salt: imageSalt })(request, none);
 
-	insertNewsletterIntoItem(item, request.promotedNewsletter)
+	insertNewsletterIntoItem(item, request.promotedNewsletter);
 
 	const newItem = {
 		...item,
